+++ Libs/DeathNotificationLib/Data/CHANGELOG.md
@@ -1,9 +1,9 @@
 # Changelog
 
-## 1773615600
+## 1774134000
 
 ### vanilla
-- Newest data as of March 16, 2026
+- 1,237,635 entries as of March 22, 2026
 
 ### tbc
-- Newest data as of March 16, 2026
+- 1,239,027 entries as of March 22, 2026